use LazyColumn instead of ProductAdapter
