+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/presentation/models/ProductsViewItem.kt
@@ -16,9 +16,4 @@ sealed class ProductsViewItem {
         val price: Double,
         val position: Int
     ) : ProductsViewItem()
-
-    data class ButtonItem(
-        val shoppingListId: String,
-        val newProductPosition: Int
-    ) : ProductsViewItem()
 }
