+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/presentation/products/mvi/ProductViewStateReducer.kt
@@ -5,7 +5,6 @@ import com.zizohanto.android.tobuy.core.ext.replaceFirst
 import com.zizohanto.android.tobuy.shopping_list.presentation.mappers.ProductModelMapper
 import com.zizohanto.android.tobuy.shopping_list.presentation.mappers.ShoppingListModelMapper
 import com.zizohanto.android.tobuy.shopping_list.presentation.mappers.ShoppingListWithProductsModelMapper
-import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem.ProductModel
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem.ShoppingListModel
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ShoppingListWithProductsModel
@@ -29,10 +28,7 @@ class ProductViewStateReducer @Inject constructor(
             is ProductViewResult.Success -> {
                 val listWithProducts: ShoppingListWithProductsModel =
                     listWithProductsMapper.mapToModel(result.listWithProducts)
-                ProductViewState.Success(
-                    listWithProducts,
-                    getSuccessViewItems(listWithProducts.shoppingList, listWithProducts.products)
-                )
+                ProductViewState.Success(listWithProducts,)
             }
             is ProductViewResult.ProductSaved -> {
                 when (previous) {
@@ -43,10 +39,7 @@ class ProductViewStateReducer @Inject constructor(
                         val savedProduct: ProductModel = productMapper.mapToModel(result.product)
                         val products: List<ProductModel> =
                             listWithProducts.products.replaceFirst(savedProduct) { it.id == savedProduct.id }
-                        ProductViewState.Success(
-                            listWithProducts.copy(products = products),
-                            getSuccessViewItems(listWithProducts.shoppingList, products)
-                        )
+                        ProductViewState.Success(listWithProducts.copy(products = products),)
                     }
                     ProductViewState.DeleteShoppingList -> TODO()
                     is ProductsViewState.Error -> ProductsViewState.Idle
@@ -61,10 +54,7 @@ class ProductViewStateReducer @Inject constructor(
                             previous.listWithProducts
                         val products: List<ProductModel> =
                             listWithProducts.products.removeFirst { it.id == result.productId }
-                        ProductViewState.Success(
-                            listWithProducts.copy(products = products),
-                            getSuccessViewItems(listWithProducts.shoppingList, products)
-                        )
+                        ProductViewState.Success(listWithProducts.copy(products = products))
                     }
                     ProductViewState.DeleteShoppingList -> TODO()
                     is ProductsViewState.Error -> TODO()
@@ -78,12 +68,7 @@ class ProductViewStateReducer @Inject constructor(
                             shoppingListModelMapper.mapToModel(result.shoppingList)
                         val listWithProducts: ShoppingListWithProductsModel =
                             previous.listWithProducts.copy(shoppingList = shoppingListModel)
-                        with(listWithProducts) {
-                            ProductViewState.Success(
-                                this,
-                                getSuccessViewItems(this.shoppingList, this.products)
-                            )
-                        }
+                        ProductViewState.Success(listWithProducts)
                     }
                     ProductViewState.DeleteShoppingList -> TODO()
                     is ProductsViewState.Error -> ProductsViewState.Idle
@@ -106,10 +91,6 @@ class ProductViewStateReducer @Inject constructor(
                     }
                     ProductViewState.Success(
                         previous.listWithProducts.copy(products = currentList),
-                        getSuccessViewItems(
-                            previous.listWithProducts.shoppingList,
-                            currentList
-                        )
                     )
                 }
                 ProductViewState.DeleteShoppingList -> TODO()
@@ -118,16 +99,4 @@ class ProductViewStateReducer @Inject constructor(
             }
         }
     }
-
-    private fun getSuccessViewItems(
-        shoppingList: ShoppingListModel,
-        products: List<ProductModel>
-    ): List<ProductsViewItem> {
-        return buildList {
-            add(shoppingList)
-            addAll(products)
-            add(ProductsViewItem.ButtonItem(shoppingList.id, products.size))
-        }
-    }
-
 }
