--- shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/ui/products/adapter/ProductViewHolder.kt
@@ -1,17 +1,0 @@
-package com.zizohanto.android.tobuy.shopping_list.ui.products.adapter
-
-import com.zizohanto.android.tobuy.shopping_list.databinding.ItemProductEditableBinding
-import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem.ProductModel
-import com.zizohanto.android.tobuy.shopping_list.presentation.views.product.RowProduct
-
-class ProductViewHolder(
-    private val binding: ItemProductEditableBinding
-) : TextChangeViewHolder(binding.root) {
-
-    fun bind(product: ProductModel, listener: ProductViewListener?) {
-        binding.products.setContent {
-            RowProduct(product, listener)
-        }
-    }
-
-}
