+++ shopping_list/src/main/java/com/zizohanto/android/tobuy/shopping_list/ui/shopping_list/adaper/SimpleProductAdapter.kt
@@ -2,6 +2,7 @@ package com.zizohanto.android.tobuy.shopping_list.ui.shopping_list.adaper
 
 import android.view.View
 import android.view.ViewGroup
+import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.zizohanto.android.tobuy.core.ext.inflate
@@ -9,7 +10,6 @@ import com.zizohanto.android.tobuy.shopping_list.R
 import com.zizohanto.android.tobuy.shopping_list.databinding.ItemProductBinding
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem
 import com.zizohanto.android.tobuy.shopping_list.presentation.models.ProductsViewItem.ProductModel
-import com.zizohanto.android.tobuy.shopping_list.ui.products.adapter.ProductAdapter.Companion.diffUtilCallback
 import com.zizohanto.android.tobuy.shopping_list.ui.shopping_list.adaper.SimpleProductAdapter.SimpleProductViewHolder
 import javax.inject.Inject
 
@@ -45,6 +45,45 @@ class SimpleProductAdapter @Inject constructor(
 
         fun bind(product: ProductModel) {
             binding.product.text = product.name
+        }
+    }
+
+    companion object {
+        val diffUtilCallback: DiffUtil.ItemCallback<ProductsViewItem>
+            get() = object : DiffUtil.ItemCallback<ProductsViewItem>() {
+                override fun areItemsTheSame(
+                    oldItem: ProductsViewItem,
+                    newItem: ProductsViewItem
+                ): Boolean {
+                    return areTheSame(oldItem, newItem)
+                }
+
+                override fun areContentsTheSame(
+                    oldItem: ProductsViewItem,
+                    newItem: ProductsViewItem
+                ): Boolean {
+                    return areTheSame(oldItem, newItem)
+                }
+            }
+
+        private fun areTheSame(
+            oldItem: ProductsViewItem,
+            newItem: ProductsViewItem
+        ): Boolean {
+            return when (oldItem) {
+                is ProductsViewItem.ShoppingListModel -> {
+                    when (newItem) {
+                        is ProductsViewItem.ShoppingListModel -> oldItem.id == newItem.id
+                        else -> false
+                    }
+                }
+                is ProductModel -> {
+                    when (newItem) {
+                        is ProductModel -> oldItem.id == newItem.id
+                        else -> false
+                    }
+                }
+            }
         }
     }
 }
