+++ shopping_list/src/main/res/layout/fragment_products.xml
@@ -14,7 +14,7 @@
         app:layout_constraintTop_toTopOf="parent"
         app:navigationIcon="@drawable/ic_ab_back_material" />
 
-    <com.zizohanto.android.tobuy.shopping_list.presentation.views.product.ProductView
+    <androidx.compose.ui.platform.ComposeView
         android:id="@+id/products_view"
         android:layout_width="0dp"
         android:layout_height="0dp"
