--- shopping_list/src/main/res/layout/item_product_editable.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.compose.ui.platform.ComposeView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/products"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content" />
