--- shopping_list/src/main/res/layout/layout_products.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.recyclerview.widget.RecyclerView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/products"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="8dp"
-    android:orientation="vertical"
-    android:paddingBottom="40dp"
-    android:clipToPadding="false"
-    android:visibility="invisible"
-    app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-    tools:context="com.zizohanto.android.tobuy.shopping_list.presentation.views.product.ProductView"
-    tools:itemCount="5"
-    tools:listitem="@layout/item_product_editable"
-    tools:visibility="visible" />
